+++ test/unit/mongo_rep_admin/code_coverage.sh
@@ -8,12 +8,12 @@ coverage erase
 
 echo ""
 echo "Running unit test modules in conjunction with coverage"
+coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/_process_std.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_mem_rep_lag.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_rep_lag.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_rep_stat.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/fetch_members.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/fetch_priority.py
-coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/fetch_rep_lag.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/get_master.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/get_optimedate.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/help_message.py