+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - \_process_std:  Process standard out formatted data.
 
 ### Changed
+- chk_mem_rep_lag:  Refactored to handle only the JSON format and convert to standard out format if requested.
 - Documentation updates.