+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [3.3.0] - 2020-07-13
+### Added
+- Added email, mongo insertion, write to file and suppression capability for the -L option in standard out format.
+- \_process_std:  Process standard out formatted data.
+
+### Changed
+- run_program:  Added repset_hosts to mongo_class.RepSet instance call.
+- chk_mem_rep_lag:  Changed dictionary to CamelCase.
+- chk_mem_rep_lag:  Refactored to handle only the JSON format and convert to standard out format if requested.
+- Documentation updates.
+
+### Removed
+- fetch_rep_lag:  No longer required.
+
+
 ## [3.2.1] - 2020-05-05
 ### Fixed
 - fetch_members:  Refactored function to use adm_cmd("replSetGetStatus") call to get the replica set data.